services/memory-controller/requirements.txt
redis>=5.0.0,<6.0.0
pydantic>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sentence-transformers>=3.0.0
